Session IDs are no longer accidently set to participant IDs
This bug was detected because we were noticing the upload feature
was not always working, and that sometimes files were being marked
as uploaded when they were not in fact uploaded. The issue was that
in the AudioPlugin, the uuid that is associated to a Participant
was being confused with the uuid that is associated with a specific
recording session. Consequently, the file names associated with the
audio recordings, rating data, and lpc were prefixed with the uuid
of the user. Since the prefix of the file names are used to determine
the session id, many different audio files have the same session id.

The consequence of this is a little bit confusing. There may be a
considerable number of audio files that are currently uploaded to the
server that are not being displayed on the interface. This is because
although the session ids are not unique, the date-time in the file name makes
the files unique, so if you go through the uploads folder / aws you will
see more audio files than there are uploads on the table displayed on the webpage.
So there is an opportunity for recovering everything that was lossed from this
bug, but it would take a little bit of work.